refactor: refactored mock data into its own file
